consolidate config step files
